fix pagination when using max_id in base url
Fixes a bug occuring when using set_max_id on the TwitterSearchOrder
object. The problem was that the base url contained a max_id
parameter, for which the pagination didn't account for, leading to two
max_id parameters in the URL of the second page. This lead to wrong
results.